% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improveCore.R
\name{anc2nwk}
\alias{anc2nwk}
\title{Converts a tree given as parent vector format to an ancestor matrix}
\usage{
anc2nwk(anc, n)
}
\arguments{
\item{anc}{A matrix. An ancestor matrix.}

\item{n}{A single integer. The number of mutations.}
}
\value{
a string of newick format tree.
}
\description{
Converts a tree given as parent vector format to an ancestor matrix
}
\examples{
p = c(11, 2, 3, 14, 14, 16, 8, 6, 9, 1, 15, 8, 10, 14, 5, 13, 17)
anc <- parentVector2anc(parents=p,n = 16)
anc2nwk(anc=anc,n=16)
}
