% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improveCore.R
\name{anc2parentVector}
\alias{anc2parentVector}
\title{Converts an ancestor matrix to a tree in parent vector format}
\usage{
anc2parentVector(anc, n)
}
\arguments{
\item{anc}{A matrix. An ancestor matrix.}

\item{n}{A single integer. The number of mutations.}
}
\value{
a integer vector; parent vector format tree.
}
\description{
Converts an ancestor matrix to a tree in parent vector format
}
\examples{
p = c(11, 2, 3, 14, 14, 16, 8, 6, 9, 1, 15, 8, 10, 14, 5, 13, 17)
anc <- parentVector2anc(parents=p,n = 16)
anc2parentVector(anc=anc,n=16)
}
