% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ancMatrixToParVector}
\alias{ancMatrixToParVector}
\title{Converts a an ancestor matrix to tree in parent vector format}
\usage{
ancMatrixToParVector(anc, n)
}
\arguments{
\item{anc}{A matrix. An ancestor matrix.}

\item{n}{A single integer. The number of mutations.}
}
\value{
a integer vector; parent vector format tree.
}
\description{
Converts a an ancestor matrix to tree in parent vector format
}
