% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFmatrixParse.R
\name{cf2treedata}
\alias{cf2treedata}
\title{Convert CF matrix to tree data}
\usage{
cf2treedata(CFmatrix_file)
}
\arguments{
\item{CFmatrix_file}{a file path to CFMatrix file.}
}
\value{
a dataframe
}
\description{
CF matrix is conflict-free matrix, tree data is a data table contains
`from`, `to` and `label` columns at least.
}
\examples{
CFmatrix_file = system.file("extdata", "ground_truth_tree.CFMatrix", package = "converTree")
cf2treedata(CFmatrix_file)
}
