% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getChildListFromParentVector}
\alias{getChildListFromParentVector}
\title{Convert a parent vector format to the list of children}
\usage{
getChildListFromParentVector(parents, n)
}
\arguments{
\item{parents}{A integer vector. Parent vector format tree.}

\item{n}{A single integer. The number of mutations.}
}
\value{
the list of children of tree.
}
\description{
Convert a parent vector format to the list of children
}
