% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFmatrixParse.R
\name{isCFM}
\alias{isCFM}
\title{Conflict matrix or Not}
\usage{
isCFM(CFmatrix_file)
}
\arguments{
\item{CFmatrix_file}{a file path to CFMatrix file.}
}
\value{
a bool value
}
\description{
A logical return, TRUE for conflict free.
}
\examples{
CFmatrix_file = system.file("extdata", "ground_truth_tree.CFMatrix", package = "converTree")
isCFM(CFmatrix_file)
}
