% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwk2xx.R
\name{nwk2anc}
\alias{nwk2anc}
\title{Convert the newick format tree to an ancestor matrix}
\usage{
nwk2anc(nwk)
}
\arguments{
\item{nwk}{the string of newick format tree.}
}
\value{
the string of newick format tree.
}
\description{
Convert the newick format tree to an ancestor matrix
}
\examples{
p = c(11, 2, 3, 14, 14, 16, 8, 6, 9, 1, 15, 8, 10, 14, 5, 13, 17)
nwk <- parentVector2nwk(parents=p,n = 16)
nwk2anc(nwk)
}
