% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx2CFmatrix.R
\name{nwk2cf}
\alias{nwk2cf}
\title{Convert nwk to CFMatrix}
\usage{
nwk2cf(nwk, filename = "test.CFMatrix")
}
\arguments{
\item{nwk}{the string of newick format tree.}

\item{filename}{file name to save CFMatrix, default is `test.CFMatrix`}
}
\value{
a dataframe and a file
}
\description{
Convert nwk to CFMatrix
}
\examples{
nwk_text="((((((((((8)7,(1)12)9)10)2)4)3,5,(((17)11)16)14)15)6)13)18;"
nwk2cf(nwk=nwk_text)
}
