% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwk2xx.R
\name{nwk2parentVector}
\alias{nwk2parentVector}
\title{Convert the newick format tree to a parent vector format}
\usage{
nwk2parentVector(nwk, n)
}
\arguments{
\item{nwk}{the string of newick format tree.}

\item{n}{A single integer. The number of mutations.}
}
\value{
a integer vector; parent vector format tree.
}
\description{
Convert the newick format tree to a parent vector format
}
\examples{
p = c(11, 2, 3, 14, 14, 16, 8, 6, 9, 1, 15, 8, 10, 14, 5, 13, 17)
nwk = parentVector2nwk(parents=p,n = 16)
nwk2parentVector(nwk,n=16)
}
