% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improveCore.R
\name{parentVector2anc}
\alias{parentVector2anc}
\title{Converts a tree given as parent vector format to an ancestor matrix}
\usage{
parentVector2anc(parents, n)
}
\arguments{
\item{parents}{A integer vector. Parent vector format tree.}

\item{n}{A single integer. The number of mutations.}
}
\value{
a logical matrix;0 and 1
}
\description{
Converts a tree given as parent vector format to an ancestor matrix
}
\examples{
p = c(11, 2, 3, 14, 14, 16, 8, 6, 9, 1, 15, 8, 10, 14, 5, 13, 17)
parentVector2anc(parents=p,n = 16)
}
