% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parentVector2ancMatrix}
\alias{parentVector2ancMatrix}
\title{Converts a tree given as parent vector format to an ancestor matrix}
\usage{
parentVector2ancMatrix(parents, n)
}
\arguments{
\item{parents}{A integer vector. Parent vector format tree.}

\item{n}{A single integer. The number of mutations.}
}
\value{
a logical matrix;TRUE and FALSE
}
\description{
Converts a tree given as parent vector format to an ancestor matrix
}
