% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improveCore.R
\name{parentVector2nwk}
\alias{parentVector2nwk}
\title{Convert a parent vector format to the newick format tree}
\usage{
parentVector2nwk(parents, n)
}
\arguments{
\item{parents}{A integer vector. Parent vector format tree.}

\item{n}{A single integer. The number of mutations.}
}
\value{
the string of newick format tree.
}
\description{
Convert a parent vector format to the newick format tree
}
\examples{
p = c(11, 2, 3, 14, 14, 16, 8, 6, 9, 1, 15, 8, 10, 14, 5, 13, 17)
parentVector2nwk(parents=p,n = 16)
}
