% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackMutaionsBranch.R
\name{stack_mutationTree}
\alias{stack_mutationTree}
\title{Stack mutations to node and make branch length to number of mutation}
\usage{
stack_mutationTree(tre)
}
\arguments{
\item{tre}{a mutation tree}
}
\value{
a tree of class "phylo"
}
\description{
stack_mutationTree is a naive function for stack mutation node in branch to
node in a mutation tree of class "phylo".
}
\examples{
library(ggtree)
nwk_text = "((((((((((8)7,(1)12)9)10)2)4)3,5,(((17)11)16)14)15)6)13)18;"
tre <- treeio::read.newick(text = nwk_text)
ggtree(tre) +
    geom_nodepoint(size = 1,color = "red",shape = 20) +
    geom_nodelab(color = "red",nudge_y = 0.1)+
    geom_nodelab(aes(label = node))+
    geom_tippoint(size = 1,color = "red",shape = 20)+
    geom_tiplab(color = "red",nudge_y = 0.1)+
    geom_tiplab(aes(label = node))
stre <- stack_mutationTree(tre)
ggtree(stre)+
  geom_nodepoint(size = 1,color = "red",shape = 20) +
  geom_nodelab(color = "red",nudge_y = 0.1,nudge_x = -0.4)+
  geom_nodelab(aes(label = node))+
  geom_tippoint(size = 1,color = "red",shape = 20)+
  geom_tiplab(color = "red",nudge_y = 0.1,nudge_x = -0.4)+
  geom_tiplab(aes(label = node))+
  theme_tree2()+
  scale_x_continuous(breaks = c(0:10), limits = c(0,12))
}
