% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xx2CFmatrix.R
\name{treedata2cf}
\alias{treedata2cf}
\title{Convert treedata to CFMatrix}
\usage{
treedata2cf(treedata, filename = "test.CFMatrix")
}
\arguments{
\item{treedata}{a treedata format dataframe}

\item{filename}{file name to save CFMatrix, default is `test.CFMatrix`}
}
\value{
a dataframe and a file
}
\description{
Convert treedata to CFMatrix, storage result to CFmatrix file and return a datafram
}
\examples{
CFmatrix_file = system.file("extdata", "ground_truth_tree.CFMatrix", package = "converTree")
tre_dat <- cf2treedata(CFmatrix_file)
cf_mat<-treedata2cf(tre_dat)
filename="test.CFMatrix"
write.table(cf_mat,file = filename,quote = FALSE,sep = "\t",row.names = FALSE)
}
