% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFmatrixParse.R
\name{treedata2gv}
\alias{treedata2gv}
\title{Convert tree data to gv file}
\usage{
treedata2gv(
  treedata,
  gvfile = "tree.gv",
  highlight = NULL,
  highcolor = "red",
  lightNodeKey = NULL,
  lightNodeColor = "red",
  is_cell = "auto_name",
  cell_prefix = "sc",
  cell_style = "filled",
  cell_shape = "box",
  cell_fill = "white",
  mut_style = "filled",
  mut_shape = "ellipse",
  mut_fill = "grey82"
)
}
\arguments{
\item{treedata}{tree data from cf2treedata}

\item{gvfile}{a file name to save result}

\item{highlight}{which mutation or cell name to highlight}

\item{highcolor}{highlight color}

\item{lightNodeKey}{mutation or cell name to target a node}

\item{lightNodeColor}{color of targeted node}

\item{is_cell}{`auto_name` or `auto_position` or a vector list of bool}

\item{cell_prefix}{if `auto_name` used in is_cell, a common prefix for cell name}

\item{cell_style}{more in [here](https://graphviz.org/docs/attrs/style/)}

\item{cell_shape}{more in [here](https://graphviz.org/doc/info/shapes.html)}

\item{cell_fill}{more in [here](https://graphviz.org/docs/attrs/fillcolor/)}

\item{mut_style}{more in [here](https://graphviz.org/docs/attrs/style/)}

\item{mut_shape}{more in [here](https://graphviz.org/doc/info/shapes.html)}

\item{mut_fill}{more in [here](https://graphviz.org/docs/attrs/fillcolor/)}
}
\value{
a file
}
\description{
[gv](https://graphviz.org/about/) file is a graph format in dot language.
}
\examples{
CFmatrix_file = system.file("extdata", "ground_truth_tree.CFMatrix", package = "converTree")
t_all = cf2treedata(CFmatrix_file)
treedata2gv(t_all,gvfile="tree.gv",highlight = c("M3","M42","M1"))
}
